feat: upgrade talisman level to uint16_t and pre-roll kill targets
- BountyTalisman.level: uint8_t → uint16_t (max level 355 overflows u8)
- BOUNTY_TALISMAN_MAX_LEVEL = 355 constant added to ioprey.hpp
- BountySlot: added optionKillTargets array (pre-rolled per option so
  kill count is stable across UI opens)
- getTalismanUpgradeCost: parameter uint8_t → uint16_t
- getTalismanBonusHundredths: new function, returns bonus in hundredths
  of a percent (2.5% base, scaling to 50% hard cap at max level)
- protocolgame: send talisman level as U16, kill count from pre-rolled
  optionKillTargets instead of re-rolling on every packet
- schema.sql + schemaTasks.sql: talisman_1..4 TINYINT → SMALLINT

Note: existing databases need:
  ALTER TABLE player_bounty MODIFY talisman_1 SMALLINT UNSIGNED NOT NULL DEFAULT 0;
  (repeat for talisman_2, talisman_3, talisman_4)

+++ src/io/functions/iologindata_load_player.cpp
@@ -925,10 +925,10 @@ void IOLoginDataLoad::loadPlayerBountyClass(const std::shared_ptr<Player> &playe
 		slot.options[0] = result->getNumber<uint16_t>("option_1");
 		slot.options[1] = result->getNumber<uint16_t>("option_2");
 		slot.options[2] = result->getNumber<uint16_t>("option_3");
-		slot.talismans[0].level = result->getNumber<uint8_t>("talisman_1");
-		slot.talismans[1].level = result->getNumber<uint8_t>("talisman_2");
-		slot.talismans[2].level = result->getNumber<uint8_t>("talisman_3");
-		slot.talismans[3].level = result->getNumber<uint8_t>("talisman_4");
+		slot.talismans[0].level = result->getNumber<uint16_t>("talisman_1");
+		slot.talismans[1].level = result->getNumber<uint16_t>("talisman_2");
+		slot.talismans[2].level = result->getNumber<uint16_t>("talisman_3");
+		slot.talismans[3].level = result->getNumber<uint16_t>("talisman_4");
 		slot.preferredSlots[0].unlocked = result->getNumber<uint8_t>("pref1_unlocked") != 0;
 		slot.preferredSlots[0].preferredRaceId = result->getNumber<uint16_t>("pref1_preferred");
 		slot.preferredSlots[0].unwantedRaceId = result->getNumber<uint16_t>("pref1_unwanted");
